requirements.txt
pyyaml
fastapi
uvicorn
pydantic>=2.0
mss
Pillow
croniter
